support: add udev rules file
Provide a udev rules file that matches on vid/pid, so that we can grant
access to the entire device.  Default openocd rules won't recognise
orbtrace as cmsis-dap, as we have the tag on the interface, not the
product, and we need to grant access to the rest of the device as well,
not just the cmsis-dap interface.

Signed-off-by: Karl Palsson <karlp@etactica.com>