msgDecoder: fix data watchpoint read handling
The least significant bit of the source address/discriminator in data
trace packets encodes whether it is a read or a write and
_handleDataRWWP properly handles this.

However, _handleHW will currently only call _handleDataRWWP when the
least significant byte is set (corresponding to a write), causing trace
packets corresponding to read operations to be discarded.

Extend the check to call _handleDataRWWP in both cases.