Correct local timestamp decoding
The bug was a simple oversight, a missing bitmask, probably due to a
copy paste error. Reformatted the three checks to be serial, so the
pattern of the code better emerges and differences are easier to spot.

Found this by emitting packages which should occur basically exactly
every 250 microseconds (160MHz, 39999 ticks), but orbuculum decoded it
to 350 (56383 ticks). The difference is 16384 ticks, so this lead me to
the parsing code.